backend/requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
python-dotenv==1.0.1
supabase==2.9.1
google-genai==0.3.0
python-multipart==0.0.12
pillow==11.0.0
requests==2.32.3
pydantic==2.9.2
